getBranchSalesById

+++ pos-system/src/main/java/com/lifepill/possystem/dto/EmployerDTO.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @Data
 public class EmployerDTO {
     private long employerId;
-    private int branchId;
+    private long branchId;
     private String employerNicName;
     private String employerFirstName;
     private String employerLastName;
