+++ pos-system/src/main/java/com/lifepill/possystem/dto/requestDTO/RegisterRequestDTO.java
@@ -15,7 +15,7 @@ import java.util.Date;
 @NoArgsConstructor
 public class RegisterRequestDTO {
     private long employerId;
-    private int branchId;
+    private long branchId;
     private String employerNicName;
     private String employerFirstName;
     private String employerLastName;
