+++ pos-system/src/main/java/com/lifepill/possystem/repo/branchRepository/BranchRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 @EnableJpaRepositories
-public interface BranchRepository extends JpaRepository<Branch,Integer> {
+public interface BranchRepository extends JpaRepository<Branch,Long> {
     boolean existsByBranchEmail(String branchEmail);
 }
